--- apollo-audit/src/main/java/com/ctrip/framework/apollo/audit/constants/ApolloAuditHttpHeader.java
@@ -1,7 +1,0 @@
-package com.ctrip.framework.apollo.audit.constants;
-
-public interface ApolloAuditHttpHeader {
-  public final String TRACE_ID = "Apollo-Audit-TraceId";
-  public final String SPAN_ID = "Apollo-Audit-SpanId";
-  public final String OPERATOR = "Apollo-Audit-Operator";
-}
