--- apollo-audit/src/main/java/com/ctrip/framework/apollo/audit/repository/ApolloAuditLogDataInfluenceRepository.java
@@ -1,8 +1,0 @@
-package com.ctrip.framework.apollo.audit.repository;
-
-import com.ctrip.framework.apollo.audit.entity.ApolloAuditLogDataInfluence;
-import org.springframework.data.repository.PagingAndSortingRepository;
-
-public interface ApolloAuditLogDataInfluenceRepository extends PagingAndSortingRepository<ApolloAuditLogDataInfluence, Long> {
-
-}
