--- apollo-audit/src/main/java/com/ctrip/framework/apollo/audit/util/ApolloAuditUtil.java
@@ -1,13 +1,0 @@
-package com.ctrip.framework.apollo.audit.util;
-
-import java.util.UUID;
-
-public class ApolloAuditUtil {
-
-  public static String generateId(){
-    return UUID.randomUUID().toString();
-  }
-
-
-
-}
