--- apollo-biz/src/main/java/com/ctrip/framework/apollo/biz/audit/ApolloAuditConfiguration.java
@@ -1,15 +1,0 @@
-package com.ctrip.framework.apollo.biz.audit;
-
-import com.ctrip.framework.apollo.audit.spi.ApolloAuditSpanService;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class ApolloAuditConfiguration {
-
-  @Bean
-  public ApolloAuditSpanService apolloAuditSpanService() {
-    return new BizAuditSpanService();
-  }
-
-}
