+++ apollo-common/pom.xml
@@ -35,9 +35,8 @@
             <artifactId>apollo-core</artifactId>
         </dependency>
         <dependency>
-            <groupId>com.ctrip.framework.apollo</groupId>
-            <artifactId>apollo-audit</artifactId>
-            <version>2.2.0-SNAPSHOT</version>
+            <groupId>com.ctrip.framework.apollo.apollo-audit</groupId>
+            <artifactId>apollo-audit-annotation</artifactId>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
