+++ apollo-portal/pom.xml
@@ -45,6 +45,10 @@
 			<artifactId>apollo-openapi</artifactId>
 		</dependency>
 		<dependency>
+			<groupId>com.ctrip.framework.apollo.apollo-audit</groupId>
+			<artifactId>apollo-audit-spring-boot-starter</artifactId>
+		</dependency>
+		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-oauth2-client</artifactId>
 		</dependency>
